Add files via upload
Here are the base Server and Client files for project 2. The server still needs some work to get the LED blinking when data is received and there are some error handling issues. I also need to modify the thread to stop along with the main program and to end once all the iterations of data are received. The client needs some updated error handling methods and needs to be further commented.