Add files via upload
I added a blinking LED for every time the server receives a new iteration of data and used threading events to set flags for the second thread to ensure a smooth exit of the server and joined the main and second thread in the main loop. Error handling was also improved to ensure that the thread didn't continue or throw errors if the Client exited or stopped transmitting data.